+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changes in git
 * Added support for multiple fallback languages!
 * Fix Django 1.4 support for using ``request.resolver_match``.
 * Fix admin delete translation view when using ``prefetch_related('translations')`` by default in the managers ``get_queryset()`` method.
+* Fix using prefetched translations in ``has_translation()`` too.
 * Return to change view after deleting a translation.